stake to module
